pyproject.toml
[project]
name = "feather-fm"
version = "0.1.0"
description = "Simple script to read songs from public Spotify 'On Repeat' playlists"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.47.0",
    "transformers>=4.57.1",
    "torch>=2.8.0",
    "librosa>=0.11.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "hdbscan>=0.8.40",
    "openai>=2.7.2",
    "rich>=14.2.0",
]
